schemaを型定義から分離
